Fix type : price_per_kg en int (col DB INTEGER, pas DECIMAL)
Erreur : 'invalid input syntax for type integer: "5000.0"'
L'app envoyait pricePerKg comme double → la DB rejetait.
Cast .toInt() avant insert.

+++ mobile/lib/services/announcement_service.dart
@@ -116,7 +116,7 @@ class AnnouncementService {
       'departure_date': departureDate.toIso8601String(),
       'arrival_date': arrivalDate?.toIso8601String(),
       'available_kg': availableKg,
-      'price_per_kg': pricePerKg,
+      'price_per_kg': pricePerKg.toInt(),
       'booked_kg': 0,
       'type': type.name,
       'status': AnnouncementStatus.pendingPayment.value,
